数据可视化/python-analysis/requirements.txt
# Web框架
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# 数据处理
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# 机器学习
scikit-learn>=1.3.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 文件处理
openpyxl==3.1.2
xlrd==2.0.1
python-multipart==0.0.6

# 工具库
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==4.6.0

# 性能优化
numba==0.57.1
cython==3.0.2

# 日志和监控
loguru==0.7.0
prometheus-client==0.17.1

# 测试
pytest==7.4.0
pytest-cov==4.1.0
